requirements.txt
petrosa
opentelemetry-distro 
opentelemetry-exporter-otlp